+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v21.10.7-1.21.10] - 2025-11-14
+## [v21.10.7-1.21.10] - 2025-11-16
+
+### Changed
+
+- Support built-in data packs that are enabled by default and can be manually disabled by players
+- Prevent clients from disconnecting for failed attachment syncs on Fabric
 
 ### Fixed